+++ src/pages/my-page/TasteProfile.tsx
@@ -5,7 +5,7 @@ import GaugeBar from '@/components/common/GaugeBar'
 import Icon from '@/components/common/Icon'
 import PackageMakeupModal from '@/components/my-page/PackageMakeupModal'
 import { ROUTE_PATHS } from '@/constants/routePaths'
-import { useTasteTestProfile } from '@/hooks/taste-test/useTasteTestProfile'
+import { useTasteTestProfile } from '@/hooks/taste-test/useTasteTest'
 import { useModal } from '@/hooks/useModal'
 import { useFeedbackProfile } from '@/hooks/user/useUser'
 import type { TasteType } from '@/types/tasteTypes'
